Struct robonomics_protocol::pubsub::Multiaddr

pub struct Multiaddr { /* private fields */ }

Representation of a Multiaddr.

Implementations

impl Multiaddr

pub fn empty() -> Multiaddr

Create a new, empty multiaddress.

pub fn with_capacity(n: usize) -> Multiaddr

Create a new, empty multiaddress with the given capacity.

pub fn len(&self) -> usize

Return the length in bytes of this multiaddress.

pub fn is_empty(&self) -> bool

Returns true if the length of this multiaddress is 0.

pub fn to_vec(&self) -> Vec<u8, Global>

Return a copy of this Multiaddr’s byte representation.

pub fn push(&mut self, p: Protocol<'_>)

Adds an already-parsed address component to the end of this multiaddr.

Examples

use multiaddr::{Multiaddr, Protocol};

let mut address: Multiaddr = "/ip4/127.0.0.1".parse().unwrap();
address.push(Protocol::Tcp(10000));
assert_eq!(address, "/ip4/127.0.0.1/tcp/10000".parse().unwrap());

pub fn pop<'a>(&mut self) -> Option<Protocol<'a>>

Pops the last Protocol of this multiaddr, or None if the multiaddr is empty.

use multiaddr::{Multiaddr, Protocol};

let mut address: Multiaddr = "/ip4/127.0.0.1/udt/sctp/5678".parse().unwrap();

assert_eq!(address.pop().unwrap(), Protocol::Sctp(5678));
assert_eq!(address.pop().unwrap(), Protocol::Udt);

pub fn with(self, p: Protocol<'_>) -> Multiaddr

Like Multiaddr::push but consumes self.

pub fn iter(&self) -> Iter<'_>

Returns the components of this multiaddress.

Example

use std::net::Ipv4Addr;
use multiaddr::{Multiaddr, Protocol};

let address: Multiaddr = "/ip4/127.0.0.1/udt/sctp/5678".parse().unwrap();

let components = address.iter().collect::<Vec<_>>();
assert_eq!(components[0], Protocol::Ip4(Ipv4Addr::new(127, 0, 0, 1)));
assert_eq!(components[1], Protocol::Udt);
assert_eq!(components[2], Protocol::Sctp(5678));

pub fn replace<'a, F>(&self, at: usize, by: F) -> Option<Multiaddr> where
F: FnOnce(&Protocol<'_>) -> Option<Protocol<'a>>,

Replace a [Protocol] at some position in this Multiaddr.

The parameter at denotes the index of the protocol at which the function by will be applied to the current protocol, returning an optional replacement.

If at is out of bounds or by does not yield a replacement value, None will be returned. Otherwise a copy of this Multiaddr with the updated Protocol at position at will be returned.

pub fn ends_with(&self, other: &Multiaddr) -> bool

Checks whether the given Multiaddr is a suffix of this Multiaddr.