Trait robonomics_protocol::pubsub::PubSub

pub trait PubSub {
    fn peer_id(&self) -> PeerId;

    fn listen(&self, address: Multiaddr) -> FutureResult<bool>;

    fn listeners(&self) -> FutureResult<Vec<Multiaddr>>;

    fn connect(&self, address: Multiaddr) -> FutureResult<bool>;

    fn subscribe<T: ToString>(&self, topic_name: &T) -> Inbox;

    fn unsubscribe<T: ToString>(&self, topic_name: &T) -> FutureResult<bool>;

    fn publish<T: ToString, M: Into<Vec<u8>>>(&self, topic_name: &T, message: M);
}

Robonomics Publisher/Subscriber.

Required methods

fn peer_id(&self) -> PeerId

Returns local peer ID.

fn listen(&self, address: Multiaddr) -> FutureResult<bool>

Listen address for incoming connections.

Returns true when successful bind and false in case of error.

fn listeners(&self) -> FutureResult<Vec<Multiaddr>>

Returns a list of node addresses.

fn connect(&self, address: Multiaddr) -> FutureResult<bool>

Connect to peer and add it into swarm.

Returns true when connected and false in case of error.

fn subscribe<T: ToString>(&self, topic_name: &T) -> Inbox

Subscribe for a topic with given name.

Returns stream of incoming messages.

fn unsubscribe<T: ToString>(&self, topic_name: &T) -> FutureResult<bool>

Unsubscribe for incoming messages from topic.

Returns true when success.

fn publish<T: ToString, M: Into<Vec<u8>>>(&self, topic_name: &T, message: M)

Publish message into the topic by name.

Loading content...